+++ app/Http/Controllers/Admin/TagController.php
@@ -6,6 +6,7 @@ use App\Http\Controllers\Controller;
 use App\Http\Requests\Admin\Tag\StoreRequest;
 use App\Http\Requests\Admin\Tag\UpdateRequest;
 use App\Models\Tag;
+use App\Models\User;
 use Illuminate\Http\Request;
 
 class TagController extends Controller
@@ -23,6 +24,7 @@ class TagController extends Controller
 
     public function store(StoreRequest $request)
     {
+        $this->authorize('create',User::class);
         $data = $request->validated();
         Tag::firstOrCreate(['name' => $data['name']]);
         return redirect()->route('admin.tags.index');
@@ -41,12 +43,14 @@ class TagController extends Controller
 
     public function update(UpdateRequest $request, Tag $tag)
     {
+        $this->authorize('update', auth()->user());
         $tag->update($request->validated());
         return redirect()->route('admin.tags.index');
     }
 
     public function destroy(Tag $tag)
     {
+        $this->authorize('delete', auth()->user());
         $tag->delete();
         return redirect()->route('admin.tags.index');
     }
